requirements.txt
asgiref==3.8.1
Django==5.0.8
djangorestframework==3.15.2
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.1
tzdata==2024.1
